+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- 2019-12-11: ActivationCoordinateActuator now provides default activation 
+              bounds equal to its control bounds.
+
 - 2019-12-10: Update the Simbody dependency to version 3.7.
 
 - 2019-12-08: Include support for OpenSim's C3DFileAdapter.